+++ esw-agent-akka/esw-agent-akka-app/src/main/scala/esw/agent/akka/app/process/cs/Coursier.scala
@@ -5,9 +5,9 @@ import esw.agent.akka.client.models.ContainerConfig
 object Coursier {
   lazy val cs: String = "cs"
 
-  def ocsApp(version: Option[String]): CoursierLaunch = CoursierLaunch("ocs-app", version)
+  def ocsApp(version: Option[String]): CoursierLaunch = CoursierLaunch("esw-ocs-app", version)
 
-  def smApp(version: Option[String]): CoursierLaunch = CoursierLaunch("sequence-manager", version)
+  def smApp(version: Option[String]): CoursierLaunch = CoursierLaunch("esw-sm-app", version)
 
   def containerApp(config: ContainerConfig): CoursierLaunch = {
     val appName = s"${config.orgName}::${config.deployModule}"