+++ esw-agent-akka/esw-agent-akka-app/src/main/scala/esw/agent/akka/app/process/cs/CoursierLaunch.scala
@@ -6,7 +6,7 @@ case class CoursierLaunch(appName: String, appVersion: Option[String]) {
   def launch(repos: List[String], mainClass: String, args: List[String]): List[String] =
     List(Coursier.cs, "launch") ::: app :: "-r" :: repos ::: "-M" :: mainClass :: "--" :: args
 
-  // cs launch --channel url://apps.json ocs-app:2.0.0 -- start seqcomp
+  // cs launch --channel url://apps.json esw-ocs-app:2.0.0 -- start seqcomp
   def launch(channel: String, args: List[String]): List[String] =
     List(Coursier.cs, "launch") ::: "--channel" :: channel :: app :: "--" :: args